+++ core/src/main/java/com/neqsoft/reso/os/Os.java
@@ -2,13 +2,26 @@ package com.neqsoft.reso.os;
 
 public class Os {
 
+  private int sdkVersion;
   private String version;
+
+  public void setSdkVersion(int sdkVersion) {
+    this.sdkVersion = sdkVersion;
+  }
+
+  public int getSdkVersion() {
+    return sdkVersion;
+  }
+
+  public void setVersion(final String version) {
+    this.version = version;
+  }
 
   public String getVersion() {
     return version;
   }
 
-  public void setVersion(String version) {
-    this.version = version;
+  public OsVersion getOsVersion() {
+    return OsVersion.get(sdkVersion);
   }
 }