+++ core/src/main/java/com/neqsoft/reso/os/OsViewModel.java
@@ -1,5 +1,7 @@
 package com.neqsoft.reso.os;
 
+import android.os.Build;
+
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 
@@ -20,7 +22,10 @@ public class OsViewModel extends ViewModel {
 
   public LiveData<Os> getOsData() {
     backgroundExecutor.execute(() -> {
-
+      Os os = new Os();
+      os.setSdkVersion(Build.VERSION.SDK_INT);
+      os.setVersion(Build.VERSION.RELEASE);
+      osData.postValue(os);
     });
     return osData;
   }